+++ src/wtforms/fields/__init__.py
@@ -1,11 +1,71 @@
-from wtforms.fields.choices import *
+from wtforms.fields.choices import RadioField
+from wtforms.fields.choices import SelectField
 from wtforms.fields.choices import SelectFieldBase
+from wtforms.fields.choices import SelectMultipleField
 from wtforms.fields.core import Field
 from wtforms.fields.core import Flags
 from wtforms.fields.core import Label
-from wtforms.fields.datetime import *
-from wtforms.fields.form import *
-from wtforms.fields.list import *
-from wtforms.fields.numeric import *
-from wtforms.fields.simple import *
+from wtforms.fields.datetime import DateField
+from wtforms.fields.datetime import DateTimeField
+from wtforms.fields.datetime import DateTimeLocalField
+from wtforms.fields.datetime import MonthField
+from wtforms.fields.datetime import TimeField
+from wtforms.fields.datetime import WeekField
+from wtforms.fields.form import FormField
+from wtforms.fields.list import FieldList
+from wtforms.fields.numeric import DecimalField
+from wtforms.fields.numeric import DecimalRangeField
+from wtforms.fields.numeric import FloatField
+from wtforms.fields.numeric import IntegerField
+from wtforms.fields.numeric import IntegerRangeField
+from wtforms.fields.simple import BooleanField
+from wtforms.fields.simple import ColorField
+from wtforms.fields.simple import EmailField
+from wtforms.fields.simple import FileField
+from wtforms.fields.simple import HiddenField
+from wtforms.fields.simple import MultipleFileField
+from wtforms.fields.simple import PasswordField
+from wtforms.fields.simple import SearchField
+from wtforms.fields.simple import StringField
+from wtforms.fields.simple import SubmitField
+from wtforms.fields.simple import TelField
+from wtforms.fields.simple import TextAreaField
+from wtforms.fields.simple import URLField
 from wtforms.utils import unset_value as _unset_value
+
+__all__ = [
+    "Field",
+    "Flags",
+    "Label",
+    "SelectField",
+    "SelectMultipleField",
+    "SelectFieldBase",
+    "RadioField",
+    "DateTimeField",
+    "DateField",
+    "TimeField",
+    "MonthField",
+    "DateTimeLocalField",
+    "WeekField",
+    "FormField",
+    "IntegerField",
+    "DecimalField",
+    "FloatField",
+    "IntegerRangeField",
+    "DecimalRangeField",
+    "BooleanField",
+    "TextAreaField",
+    "PasswordField",
+    "FileField",
+    "MultipleFileField",
+    "HiddenField",
+    "SearchField",
+    "SubmitField",
+    "StringField",
+    "TelField",
+    "URLField",
+    "EmailField",
+    "ColorField",
+    "FieldList",
+    "_unset_value",
+]