+++ src/wtforms/fields/choices.py
@@ -30,7 +30,7 @@ class SelectFieldBase(Field):
     def iter_choices(self):
         """
         Provides data for choice widget rendering. Must return a sequence or
-        iterable of (value, label, selected) tuples.
+        iterable of (value, label, selected, render_kw) tuples.
         """
         raise NotImplementedError()
 
@@ -49,8 +49,14 @@ class SelectFieldBase(Field):
             _form=None,
             _meta=self.meta,
         )
-        for i, (value, label, checked) in enumerate(self.iter_choices()):
-            opt = self._Option(label=label, id="%s-%d" % (self.id, i), **opts)
+        for i, choice in enumerate(self.iter_choices()):
+            if len(choice) == 4:
+                value, label, checked, render_kw = choice
+            else:
+                value, label, checked = choice
+                render_kw = {}
+
+            opt = self._Option(label=label, id=f"{self.id}-{i}", **opts, **render_kw)
             opt.process(None, value)
             opt.checked = checked
             yield opt
@@ -106,14 +112,16 @@ class SelectField(SelectFieldBase):
         if not choices:
             _choices = []
 
-        elif isinstance(choices[0], (list, tuple)):
+        elif isinstance(choices[0], list | tuple):
             _choices = choices
 
         else:
-            _choices = zip(choices, choices)
+            _choices = zip(choices, choices, strict=False)
 
-        for value, label in _choices:
-            yield (value, label, self.coerce(value) == self.data)
+        for value, label, *other_args in _choices:
+            selected = self.coerce(value) == self.data
+            render_kw = other_args[0] if len(other_args) else {}
+            yield (value, label, selected, render_kw)
 
     def process_data(self, value):
         try:
@@ -132,13 +140,13 @@ class SelectField(SelectFieldBase):
             raise ValueError(self.gettext("Invalid Choice: could not coerce.")) from exc
 
     def pre_validate(self, form):
-        if self.choices is None:
-            raise TypeError(self.gettext("Choices cannot be None."))
-
         if not self.validate_choice:
             return
 
-        for _, _, match in self.iter_choices():
+        if self.choices is None:
+            raise TypeError(self.gettext("Choices cannot be None."))
+
+        for _, _, match, *_ in self.iter_choices():
             if match:
                 break
         else:
@@ -148,24 +156,26 @@ class SelectField(SelectFieldBase):
 class SelectMultipleField(SelectField):
     """
     No different from a normal select field, except this one can take (and
-    validate) multiple choices.  You'll need to specify the HTML `size`
-    attribute to the select field when rendering.
+    validate) multiple choices. You'll need to specify the HTML
+    :mdn-attr:`size` attribute on the :mdn-tag:`select` field when rendering.
     """
 
     widget = widgets.Select(multiple=True)
 
     def _choices_generator(self, choices):
-        if choices:
-            if isinstance(choices[0], (list, tuple)):
-                _choices = choices
-            else:
-                _choices = zip(choices, choices)
-        else:
+        if not choices:
             _choices = []
 
-        for value, label in _choices:
+        elif isinstance(choices[0], list | tuple):
+            _choices = choices
+
+        else:
+            _choices = zip(choices, choices, strict=False)
+
+        for value, label, *other_args in _choices:
             selected = self.data is not None and self.coerce(value) in self.data
-            yield (value, label, selected)
+            render_kw = other_args[0] if len(other_args) else {}
+            yield (value, label, selected, render_kw)
 
     def process_data(self, value):
         try:
@@ -184,15 +194,17 @@ class SelectMultipleField(SelectField):
             ) from exc
 
     def pre_validate(self, form):
-        if self.choices is None:
-            raise TypeError(self.gettext("Choices cannot be None."))
-
         if not self.validate_choice or not self.data:
             return
 
-        acceptable = {c[0] for c in self.iter_choices()}
-        if any(d not in acceptable for d in self.data):
-            unacceptable = [str(d) for d in set(self.data) - acceptable]
+        if self.choices is None:
+            raise TypeError(self.gettext("Choices cannot be None."))
+
+        acceptable = [self.coerce(choice[0]) for choice in self.iter_choices()]
+        if any(data not in acceptable for data in self.data):
+            unacceptable = [
+                str(data) for data in set(self.data) if data not in acceptable
+            ]
             raise ValidationError(
                 self.ngettext(
                     "'%(value)s' is not a valid choice for this field.",
@@ -205,7 +217,7 @@ class SelectMultipleField(SelectField):
 
 class RadioField(SelectField):
     """
-    Like a SelectField, except displays a list of radio buttons.
+    Like a SelectField, except displays a list of :mdn-input:`radio` buttons.
 
     Iterating the field will produce subfields (each containing a label as
     well) in order to allow custom rendering of the individual radio fields.
@@ -213,3 +225,7 @@ class RadioField(SelectField):
 
     widget = widgets.ListWidget(prefix_label=False)
     option_widget = widgets.RadioInput()
+
+    def __init__(self, label=None, validators=None, **kwargs):
+        super().__init__(label, validators, **kwargs)
+        self.label.field_id = False