+++ src/wtforms/fields/list.py
@@ -1,9 +1,10 @@
 import itertools
 
+from wtforms.utils import unset_value
+
 from .. import widgets
 from .core import Field
 from .core import UnboundField
-from wtforms.utils import unset_value
 
 __all__ = ("FieldList",)
 
@@ -49,9 +50,9 @@ class FieldList(Field):
                 "FieldList does not accept any filters. Instead, define"
                 " them on the enclosed field."
             )
-        assert isinstance(
-            unbound_field, UnboundField
-        ), "Field must be unbound, not a field class"
+        assert isinstance(unbound_field, UnboundField), (
+            "Field must be unbound, not a field class"
+        )
         self.unbound_field = unbound_field
         self.min_entries = min_entries
         self.max_entries = max_entries
@@ -67,6 +68,7 @@ class FieldList(Field):
                 " them on the enclosed field."
             )
 
+        self.last_index = -1
         self.entries = []
         if data is unset_value or not data:
             try:
@@ -143,7 +145,7 @@ class FieldList(Field):
         candidates = itertools.chain(ivalues, itertools.repeat(None))
         _fake = type("_fake", (object,), {})
         output = []
-        for field, data in zip(self.entries, candidates):
+        for field, data in zip(self.entries, candidates, strict=False):
             fake_obj = _fake()
             fake_obj.data = data
             field.populate_obj(fake_obj, "data")
@@ -152,22 +154,22 @@ class FieldList(Field):
         setattr(obj, name, output)
 
     def _add_entry(self, formdata=None, data=unset_value, index=None):
-        assert (
-            not self.max_entries or len(self.entries) < self.max_entries
-        ), "You cannot have more than max_entries entries in this FieldList"
+        assert not self.max_entries or len(self.entries) < self.max_entries, (
+            "You cannot have more than max_entries entries in this FieldList"
+        )
         if index is None:
             index = self.last_index + 1
         self.last_index = index
         name = f"{self.short_name}{self._separator}{index}"
         id = f"{self.id}{self._separator}{index}"
-        field = self.unbound_field.bind(
-            form=None,
+        options = dict(
             name=name,
             prefix=self._prefix,
             id=id,
             _meta=self.meta,
             translations=self._translations,
         )
+        field = self.meta.bind_field(None, self.unbound_field, options)
         field.process(formdata, data)
         self.entries.append(field)
         return field