+++ src/wtforms/fields/numeric.py
@@ -144,7 +144,7 @@ class DecimalField(LocaleAwareNumberField):
         if not hasattr(self.data, "quantize"):
             # If for some reason, data is a float or int, then format
             # as we would for floats using string formatting.
-            format = "%%0.%df" % self.places
+            format = f"%.{self.places}f"
             return format % self.data
 
         exp = decimal.Decimal(".1") ** self.places
@@ -170,11 +170,14 @@ class DecimalField(LocaleAwareNumberField):
 
 class FloatField(Field):
     """
-    A text field, except all input is coerced to an float.  Erroneous input
-    is ignored and will not be accepted as a value.
+    A field that stores floating-point values.
+
+    By default, this renders as an ``<input type="number">`` with
+    ``step="any"`` to allow browser input of any floating-point value.
+    Erroneous input is ignored and will not be accepted as a value.
     """
 
-    widget = widgets.TextInput()
+    widget = widgets.NumberInput(step="any")
 
     def __init__(self, label=None, validators=None, **kwargs):
         super().__init__(label, validators, **kwargs)
@@ -199,7 +202,7 @@ class FloatField(Field):
 
 class IntegerRangeField(IntegerField):
     """
-    Represents an ``<input type="range">``.
+    Represents an :mdn-input:`range`.
     """
 
     widget = widgets.RangeInput()
@@ -207,7 +210,7 @@ class IntegerRangeField(IntegerField):
 
 class DecimalRangeField(DecimalField):
     """
-    Represents an ``<input type="range">``.
+    Represents an :mdn-input:`range`.
     """
 
     widget = widgets.RangeInput(step="any")