+++ tests/common.py
@@ -1,6 +1,6 @@
 class DummyPostData(dict):
     def getlist(self, key):
         v = self[key]
-        if not isinstance(v, (list, tuple)):
+        if not isinstance(v, list | tuple):
             v = [v]
         return v