+++ tests/fields/test_html5.py
@@ -1,9 +1,9 @@
+import warnings
 from datetime import date
 from datetime import datetime
 from decimal import Decimal
 
 from tests.common import DummyPostData
-
 from wtforms import validators
 from wtforms.fields import DateField
 from wtforms.fields import DateTimeField
@@ -11,6 +11,7 @@ from wtforms.fields import DateTimeLocalField
 from wtforms.fields import DecimalField
 from wtforms.fields import DecimalRangeField
 from wtforms.fields import EmailField
+from wtforms.fields import FloatField
 from wtforms.fields import IntegerField
 from wtforms.fields import IntegerRangeField
 from wtforms.fields import PasswordField
@@ -23,10 +24,6 @@ from wtforms.form import Form
 from wtforms.utils import unset_value
 
 
-def make_form(name="F", **fields):
-    return type(str(name), (Form,), fields)
-
-
 class F(Form):
     search = SearchField()
     telephone = TelField()
@@ -36,6 +33,7 @@ class F(Form):
     date = DateField()
     dt_local = DateTimeLocalField()
     integer = IntegerField()
+    float = FloatField()
     decimal = DecimalField()
     int_range = IntegerRangeField()
     decimal_range = DecimalRangeField()
@@ -45,8 +43,8 @@ def _build_value(key, form_input, expected_html, data=unset_value):
     if data is unset_value:
         data = form_input
     if expected_html.startswith("type="):
-        expected_html = '<input id="{}" name="{}" {} value="{}">'.format(
-            key, key, expected_html, form_input
+        expected_html = (
+            f'<input id="{key}" name="{key}" {expected_html} value="{form_input}">'
         )
     return {
         "key": key,
@@ -83,10 +81,15 @@ def test_simple():
             42,
         ),
         b(
+            "float",
+            "1.234",
+            '<input id="float" name="float" step="any" type="number" value="1.234">',
+            1.234,
+        ),
+        b(
             "decimal",
             "43.5",
-            '<input id="decimal" name="decimal" '
-            'step="any" type="number" value="43.5">',
+            '<input id="decimal" name="decimal" step="any" type="number" value="43.5">',
             Decimal("43.5"),
         ),
         b(
@@ -109,16 +112,19 @@ def test_simple():
         formdata[item["key"]] = item["form_input"]
         kw[item["key"]] = item["data"]
 
-    form = F(formdata)
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", DeprecationWarning)
+        form = F(formdata)
     for item in VALUES:
         field = form[item["key"]]
         render_value = field()
-        if render_value != item["expected_html"]:
-            tmpl = "Field {key} render mismatch: {render_value!r} != {expected_html!r}"
-            raise AssertionError(tmpl.format(render_value=render_value, **item))
-        if field.data != item["data"]:
-            tmpl = "Field {key} data mismatch: {field.data!r} != {data!r}"
-            raise AssertionError(tmpl.format(field=field, **item))
+        assert render_value == item["expected_html"], (
+            f"Field {item['key']} render mismatch: "
+        )
+        "{render_value} != {item['expected_html']}"
+        assert field.data == item["data"], (
+            "Field {item['key']} data mismatch: {field.data} != {item['data']}"
+        )
 
 
 class G(Form):